Update wxLua to 3.0.0.6 + compilation patch.
Now using the fork at https://github.com/pkulchenko/wxlua as the primary
project has not issued a release in nearly 7 years. The fork has made
many updates, including:

* Better compatibility with newer versions of wxWidgets
* Fixed platform inconsistencies with bit library, which led to
  incorrect md5 calculations on Mac (see #61, pkulchenko/wxlua#43). The
  md5 sums have been renegerated and are now correct on all platforms.

To fix builds with wxWidgets < 3.1.2, we also include the fix for
pkulchenko/wxlua#46.


$ rsync -r --delete \
    --exclude="premake4.lua" \
    --exclude="wxluasetup_xword.h" \
    --exclude="Makefile*" \
    --exclude="CMakeLists.txt" \
    wxlua-3.0.0.6/wxLua/modules/wxbind \
    wxlua-3.0.0.6/wxLua/modules/wxlua \
    xword/lua/
$ rsync -r --delete \
    wxlua-3.0.0.6/wxLua/art \
    xword/lua/wxlua/
$ cp wxlua-3.0.0.6/wxLua/bindings/genwxbind.lua xword/src/xwordbind/

Along with manual edits to wxluasetup_xword.h to sync with the latest
set of config options in wxluasetup.h (i.e. disable new entries).

Regenerated bindings in src/xwordbind/.

Disabled webview in src/xwordlua.hpp and replaced obsolete reference to
wxBG_STYLE_CUSTOM in sizedtext.lua.

+++ lua/wxlua/wxllua.cpp
@@ -199,7 +199,7 @@ void LUACALL wxlua_argerrormsg(lua_State *L, const wxString& msg_)
 
     wxString msg;
     msg.Printf(wxT("%s\nFunction called: '%s'%s"), msg_.c_str(), argMsg.c_str(), funcArgs.c_str());
-    wxlua_error(L, msg);
+    wxlua_error(L, msg.c_str());
 }
 
 void* LUACALL wxlua_touserdata(lua_State *L, int stack_idx, bool null_ptr /*= false*/)
@@ -389,7 +389,7 @@ bool LUACALL wxluaO_deletegcobject(lua_State *L, int stack_idx, int flags)
     void* udata   = lua_touserdata(L, stack_idx);
     void* obj_ptr = wxlua_touserdata(L, stack_idx, true); // clear lua userdata's ptr
 
-    //if (obj_ptr == NULL) return false; // can happen
+    if (obj_ptr == NULL) return false; // can happen
 
     bool delete_all = WXLUA_HASBIT(flags, WXLUA_DELETE_OBJECT_ALL);
 
@@ -1068,7 +1068,8 @@ void* LUACALL wxluaT_getuserdatatype(lua_State* L, int stack_idx, int wxl_type)
         // pointer to the class object on the stack. We need to shift the
         // pointer by the number of bytes in wxLuaBindClass::baseclass_vtable_offsets
         // so that when it is casted to the base class we don't segfault.
-        long int o = (long int)wxlua_touserdata(L, stack_idx, false);
+        // Using 'long long' for 32 and 64 bit and compatibility with older compilers that don't have uintptr_t.
+        unsigned long long o = (unsigned long long)wxlua_touserdata(L, stack_idx, false);
 
         if (wxlClass->baseclass_wxluatypes)
         {
